Multichannel download now working, along with various other changes:
1. Multichannel debugged to the point where it now works without issue,
though further testing will take place to verify.

2. End Game mode has been implemented to prevent issues such as hanging
at the end of the download.

3. Print statements have been swapped out with logging statements.
These write to the console in colour and to a file in a new Logs folder
for further analysis.

4. An optimisation "piecesMark" has been introduced in the RarestFirst
class.  This tracks sequentially up the pieces array so long as each
piece it checks has been obtained.  Certain operations on the pieces
array can base their start point off of piecesMark instead of the start
of the array to reduce the number of array accesses required.