Changes made to the way pieces are sorted:
A flaw resulted in pieces with high availability being selected first.
This has now been reversed.  When collected a piece's frequency is now
set to INTEGER.MIN_VALUE.  This sets it to the front of the array and
allows piecesMark to move past it when possible, resulting in the first
part of the array consisting of pieces collected and the last part
consisting of pieces to collect.

The pieces array is sorted at intervals to allow for the re-organisation
of pieces collected and pieces waiting to be collected.  This is done
for every 10% of the file that is downloaded, excluding 100%.